Move timing out of core.util
